Refactor file, progress on hash to field
